dbft: rename package from `clique` to `dbft`
