+++ consensus/dbft/dbft.go
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-// Package clique implements the proof-of-authority consensus engine.
-package clique
+// Package dbft implements the proof-of-authority consensus engine.
+package dbft
 
 import (
 	"bytes"
